refactor/relatorios: extrai componentes compartilhados
