+++ front-end/src/pages/reports/ReportsOperational.tsx
@@ -1,40 +1,7 @@
 import React from 'react';
 import { Filter, Route, Truck } from 'lucide-react';
 import { useReportsData } from './useReportsData';
-
-function Panel({ title, children }: { title: string; children: React.ReactNode }) {
-  return (
-    <section className="bg-surface-container-lowest rounded-3xl border border-outline-variant p-6">
-      <h3 className="mb-5 text-xl font-bold text-on-surface">{title}</h3>
-      {children}
-    </section>
-  );
-}
-
-function MetricBox({ label, value, icon: Icon, highlight }: { label: string; value: string; icon: React.ElementType; highlight?: boolean }) {
-  return (
-    <div className={`rounded-3xl border p-6 shadow-sm ${highlight ? 'border-primary/20 bg-primary-container/20' : 'border-outline-variant bg-surface-container-lowest'}`}>
-      <div className="mb-3 flex items-center justify-between">
-        <p className="text-sm font-medium text-on-surface-variant">{label}</p>
-        <Icon className={`h-5 w-5 ${highlight ? 'text-primary' : 'text-on-surface-variant'}`} />
-      </div>
-      <p className="text-3xl font-black text-on-surface">{value}</p>
-    </div>
-  );
-}
-
-function EmptyText({ text }: { text: string }) {
-  return <p className="text-sm text-on-surface-variant">{text}</p>;
-}
-
-function MapMarkerIcon(props: React.ComponentProps<'svg'>) {
-  return (
-    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" {...props}>
-      <path d="M12 22s7-4.35 7-11a7 7 0 1 0-14 0c0 6.65 7 11 7 11Z" />
-      <circle cx="12" cy="11" r="3" />
-    </svg>
-  );
-}
+import { EmptyText, MapMarkerIcon, MetricBox, Panel } from './ReportsSharedComponents';
 
 type ReportsOperationalProps = {
   data: ReturnType<typeof useReportsData>;
